requirements.txt
streamlit>=1.31.0
groq>=0.4.2
python-dotenv>=1.0.0
PyPDF2>=3.0.0
numpy>=1.24.0
langchain>=0.1.0
pypdf>=3.17.1
chromadb>=0.4.22